- [tile] return vector by const ref instead of copy
- Updated call to getLength
- Fix out-of-bounds access in Platform::getCPUName
- Use std::deque for queues
- Fixed window returning from fullscreen

+++ src/framework/stdext/time.h
@@ -27,23 +27,22 @@
 
 namespace stdext {
 
-ticks_t time();
-ticks_t millis();
-ticks_t micros();
-void millisleep(size_t ms);
-void microsleep(size_t us);
+    ticks_t time();
+    ticks_t millis();
+    ticks_t micros();
+    void millisleep(size_t ms);
+    void microsleep(size_t us);
 
-struct timer {
-public:
-    timer() { restart(); }
-    float elapsed_seconds() { return static_cast<float>((stdext::micros() - m_start) / 1000000.0); }
-    ticks_t elapsed_millis() { return (stdext::micros() - m_start)/1000; }
-    ticks_t elapsed_micros() { return stdext::micros() - m_start; }
-    void restart() { m_start = stdext::micros(); }
-private:
-    ticks_t m_start;
-};
-
+    struct timer {
+    public:
+        timer() { restart(); }
+        float elapsed_seconds() { return (stdext::micros() - m_start) / 1000000.f; }
+        ticks_t elapsed_millis() { return (stdext::micros() - m_start) / 1000; }
+        ticks_t elapsed_micros() { return stdext::micros() - m_start; }
+        void restart() { m_start = stdext::micros(); }
+    private:
+        ticks_t m_start;
+    };
 }
 
 #endif
